ensure root's user.name and user.email are configured
On occasion, I've noticed some weirdness when root does not have a git user
and email address defined.  This adds configuration to ensure that
~root/.gitconfig has a basic user and email defined, e.g.

    user.name   root
    user.email  root@${hostname}

Note that this depends on puppetlabs-stdlib's custom fact of $root_dir being
available...  As the stdlib module is a pretty common requirement, I'm not
terribly worried about it here, but we could find an alternate way if
essential.

Another approach would be to create, say, a /etc/etckeeper/init.d/41-git-name,
and have that set the config from inside a new repository.  This would solve
the problem going forward, but not so much for any current systems.